backend/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
PyJWT
bcrypt
python-dotenv
Flask-Cors
pandas
openpyxl